refactor move logic into widget
